E_gzamin_django/requirements.txt
Django==3.0.7
psycopg2==2.8.5
djangorestframework==3.11.0
drf-extensions==0.6.0